Website/requirements.txt
flask==3.0.2
tensorflow==2.10.1
selenium==4.18.1
webdriver-manager==4.0.1
Pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0 